backend-services/order_service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pymysql
pika==1.3.2
cryptography
django
djangorestframework